+++ src/main/java/com/cafe/website/entity/ProductDiscount.java
@@ -17,7 +17,7 @@ public class ProductDiscount extends BaseEntity {
 	private Integer percent;
 	private String name;
 
-	public ProductDiscount(int id, int status, String createdAt, String updatedAt, Product product, Long expiryDate,
+	public ProductDiscount(Long id, int status, String createdAt, String updatedAt, Product product, Long expiryDate,
 			Integer percent, String name) {
 		super(id, status, createdAt, updatedAt);
 		this.product = product;
