+++ src/main/java/com/cafe/website/entity/Review.java
@@ -43,7 +43,7 @@ public class Review extends BaseEntity {
 	@Column(unique = true)
 	private List<Favorite> favorites = new ArrayList<>();
 
-	public Review(int id, int status, String createdAt, String updatedAt, User user, Product product, Rating rating,
+	public Review(Long id, int status, String createdAt, String updatedAt, User user, Product product, Rating rating,
 			List<Comment> comments, List<Image> listImages, String name, List<Favorite> favorites) {
 		super(id, status, createdAt, updatedAt);
 		this.user = user;
